+++ packages/core/development.env
@@ -29,8 +29,8 @@ FFPROBE_PATH=""
 # ffplay
 FFPLAY_PATH=""
 # 这里请勿修改
-RUNTIME="node"
+RUNTIME="tsx"
 # 运行模式
-NODE_ENV="production"
+NODE_ENV="development"
 # tsx watch
 TSX_WATCH="false"
