+++ packages/core/package.json
@@ -53,12 +53,12 @@
     "build:dev": "pnpm build && node cli/build.js -dev",
     "build:module": "tsup --tsconfig tsconfig.module.json --config tsup.modules.ts && node cli/module.js",
     "build:prod": "pnpm build:module && pnpm build:cli && pnpm build && node cli/build.js",
+    "bun:app": "bun dist/index.js",
+    "bun:dev": "cross-env EBV_FILE=\"development.env\" bun --import tsx src/index.ts",
     "cli": "node dist/cli/index.cjs",
     "debug": "node dist/index.js --debug",
-    "dev": "tsx src/index.ts",
+    "dev": "cross-env EBV_FILE=\"development.env\" node --import tsx src/index.ts",
     "dev:cli": "tsx exports/cli/index.ts",
-    "dev:w": "tsx watch --include \"./src/**/*.ts\" src/index.ts",
-    "dev:watch": "pnpm dev:w",
     "pr": "node cli/pr.js all",
     "pub": "npm publish --access public",
     "pub-beta": "npm publish --access public --tag beta",
@@ -88,8 +88,9 @@
     "yaml": "2.7.0"
   },
   "devDependencies": {
+    "@karinjs/node-pty": "^1.0.4",
     "@types/jsonwebtoken": "^9.0.9",
-    "@karinjs/node-pty": "^1.0.4"
+    "cross-env": "^7.0.3"
   },
   "peerDependencies": {
     "node-karin": "workspace:*"
