+++ packages/core/src/server/system/manage.ts
@@ -17,6 +17,6 @@ export const restartRouter: RequestHandler = async (_req, res) => {
 export const exitRouter: RequestHandler = async (_req, res) => {
   logger.mark('收到退出请求，正在退出...')
   createSuccessResponse(res, null, '退出指令发送成功')
-  const { processExit } = await import('@/core/internal/process')
+  const { processExit } = await import('@/service/process/exit')
   await processExit(0)
 }
