+++ packages/core/tsup.cli.ts
@@ -1,21 +1,11 @@
-import fs from 'node:fs'
-import { URL } from 'node:url'
 import { defineConfig } from 'tsup'
-
-const pkg = JSON.parse(fs.readFileSync(new URL('package.json', import.meta.url), 'utf-8'))
-const noExternal = ['jsonwebtoken', 'systeminformation', 'commander']
-const external = Object.keys(pkg.dependencies).filter(dep => !noExternal.includes(dep))
+import { options } from './tsup.base'
 
 export default defineConfig({
+  ...options,
   entry: ['exports/cli/index.ts'],
   format: ['cjs', 'esm'],
-  target: 'node16',
-  splitting: false,
-  sourcemap: false,
   clean: true,
-  dts: true,
   outDir: 'dist/cli',
   treeshake: true,
-  external,
-  noExternal,
 })
